+++ examples/bop.rb
@@ -9,7 +9,7 @@ ROOM = ENV['ROOM']          # 'xxxxxxxxxxxxxxxxxxxxxxxx'
 TT.run(EMAIL, PASSWORD, :room => ROOM) do
   on :user_spoke do |message|
     if message.content =~ /bop/
-      client.room.current_song.vote
+      room.current_song.vote
     end
   end
 end