+++ examples/chat_bot.rb
@@ -10,7 +10,7 @@ TT.run(EMAIL, PASSWORD, :room => ROOM) do
   on :user_spoke do |message|
     # Respond to "/hello" command
     if (message.content =~ /^\/hello$/)
-      client.user.say("Hey! How are you @#{message.sender.name}?")
+      room.say("Hey! How are you @#{message.sender.name}?")
     end
   end
 end